fix restore hidden shared libraries

+++ ntp-arm-musl.sh
@@ -937,7 +937,10 @@ finalize_build() {
 
 # temporarily hide shared libraries (.so) to force cmake to use the static ones (.a)
 hide_shared_libraries() {
-    mv -f "${PREFIX}/lib_hidden/"* "${PREFIX}/lib/" || true
+    if [ -d "${PREFIX}/lib_hidden" ]; then
+        mv -f "${PREFIX}/lib_hidden/"* "${PREFIX}/lib/" || true
+        rmdir "${PREFIX}/lib_hidden" || true
+    fi
     mkdir -p "${PREFIX}/lib_hidden" || true
     mv -f "${PREFIX}/lib/"*".so"* "${PREFIX}/lib_hidden/" || true
     mv -f "${PREFIX}/lib_hidden/libcc1."* "${PREFIX}/lib/" || true
@@ -946,8 +949,10 @@ hide_shared_libraries() {
 
 # restore the hidden shared libraries
 restore_shared_libraries() {
-    mv -f "${PREFIX}/lib_hidden/"* "${PREFIX}/lib/" || true
-    rmdir "${PREFIX}/lib_hidden" || true
+    if [ -d "${PREFIX}/lib_hidden" ]; then
+        mv -f "${PREFIX}/lib_hidden/"* "${PREFIX}/lib/" || true
+        rmdir "${PREFIX}/lib_hidden" || true
+    fi
     return 0
 }
 
@@ -1071,6 +1076,9 @@ if [ ! -d "${CROSSBUILD_DIR}" ]; then
     verify_hash "${PKG_SOURCE_PATH}" "${PKG_HASH}"
     unpack_archive "${PKG_SOURCE_PATH}" "${CROSSBUILD_DIR}"
 fi
+
+# restore the hidden shared libraries, if they were not previously restored
+restore_shared_libraries
 
 # Check for required toolchain tools
 if [ ! -x "${CROSSBUILD_DIR}/bin/${TARGET}-gcc" ]; then
